Moved MQTT state updated from main to IrrigationController
This actually implements state updates in the first place. Moved the
unused stuff from main away.

It also handles the MAC address stuff more locally, where needed.

The PowerManager delivers the batteryVoltage now as an uint32_t, which
is more lightweight when debug output is disabled.

+++ main/include/powerManager.h
@@ -19,7 +19,6 @@
 
 #include "user_config.h"
 
-
 class PowerManager
 {
 private:
@@ -44,7 +43,7 @@ private:
 public:
     PowerManager(void);
 
-    float getSupplyVoltageMilli(void);
+    uint32_t getSupplyVoltageMilli(void);
 
     void setPeripheralEnable(bool en);
     bool getPeripheralEnable(void);
